+++ src/gui/plot.hpp
@@ -8,4 +8,4 @@
 #include <vector>
 #include <gui/color.hpp>
 
-void plot_2d_path(const std::string& label, const std::vector<Eigen::Vector2d>& positions, const Color& color = Color::Auto(), float weight = -1.0f);
+void plot_2d_path(const std::string& label, const std::vector<Eigen::Vector2f>& positions, const Color& color = Color::Auto(), float weight = -1.0f);
